# server/openapi.json
{
  "components": {
    "schemas": {
...
      "CatalogModel": {
        "description": "A catalog entry: the OpenAI model fields plus the operator's favorite flag\nand the custom-model marker Node adds in `MergeCustomModels`.",
        "properties": {
          "custom": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "favorite": {
            "type": "boolean"
          },
          "id": {
            "type": "string"
          },
          "object": {
            "type": "string"
          },
          "owned_by": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "object",
          "owned_by",
          "favorite"
        ],
        "title": "CatalogModel",
        "type": "object"
      },
      "CatalogResponse": {
        "description": "`GET /v1/providers/catalog`.",
        "properties": {
          "categories": {
            "$ref": "#/components/schemas/GroupedCatalog"
          },
          "total": {
            "format": "uint",
            "minimum": 0,
            "type": "integer"
          }
        },
        "required": [
          "total",
          "categories"
        ],
        "title": "CatalogResponse",
        "type": "object"
      },
...
      "GroupedCatalog": {
        "description": "The four fixed provider groups.",
        "properties": {
          "api_key": {
            "items": {
              "$ref": "#/components/schemas/ProviderEntry"
            },
            "type": "array"
          },
          "custom_provider": {
            "items": {
              "$ref": "#/components/schemas/ProviderEntry"
            },
            "type": "array"
          },
          "free_tier": {
            "items": {
              "$ref": "#/components/schemas/ProviderEntry"
            },
            "type": "array"
          },
          "oauth": {
            "items": {
              "$ref": "#/components/schemas/ProviderEntry"
            },
            "type": "array"
          }
        },
        "required": [
          "oauth",
          "free_tier",
          "api_key",
          "custom_provider"
        ],
        "type": "object"
      },
...
      "ModelListResponse": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/CatalogModel"
            },
            "type": "array"
          },
          "object": {
            "type": "string"
          }
        },
        "required": [
          "object",
          "data"
        ],
        "title": "ModelListResponse",
        "type": "object"
      },
...
      "ProviderConnectionView": {
        "description": "A connection as the detail response reports it. There is deliberately no\ncredential field: a stored API key never leaves the database.",
        "properties": {
          "alias": {
            "type": [
              "string",
              "null"
            ]
          },
          "base_url": {
            "type": [
              "string",
              "null"
            ]
          },
          "category": {
            "type": "string"
          },
          "created_at": {
            "format": "int64",
            "type": "integer"
          },
          "enabled": {
            "type": "boolean"
          },
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "protocol": {
            "type": "string"
          },
          "provider_id": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "provider_id",
          "name",
          "category",
          "protocol",
          "enabled",
          "created_at"
        ],
        "type": "object"
      },
      "ProviderEntry": {
        "description": "A provider as the list, catalog, and detail routes describe it.",
        "properties": {
          "category": {
            "type": "string"
          },
          "connections": {
            "description": "Detail only; the list and catalog omit it like Node does.",
            "items": {
              "$ref": "#/components/schemas/ProviderConnectionView"
            },
            "type": [
              "array",
              "null"
            ]
          },
          "default_base_url": {
            "type": "string"
          },
          "enabled": {
            "type": "boolean"
          },
          "id": {
            "type": "string"
          },
          "models": {
            "items": {
              "$ref": "#/components/schemas/ProviderModel"
            },
            "type": "array"
          },
          "name": {
            "type": "string"
          },
          "protocol": {
            "$ref": "#/components/schemas/ProviderProtocol"
          },
          "requires_api_key": {
            "type": "boolean"
          },
          "requires_oauth": {
            "type": "boolean"
          },
          "round_robin": {
            "description": "Whether requests rotate across this provider's accounts.",
            "type": "boolean"
          },
          "status": {
            "$ref": "#/components/schemas/ProviderStatus"
          },
          "supports_custom_url": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "name",
          "category",
          "protocol",
          "default_base_url",
          "requires_api_key",
          "requires_oauth",
          "supports_custom_url",
          "enabled",
          "round_robin",
          "status",
          "models"
        ],
        "type": "object"
      },
      "ProviderListResponse": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/ProviderEntry"
            },
            "type": "array"
          },
          "object": {
            "type": "string"
          }
        },
        "required": [
          "object",
          "data"
        ],
        "title": "ProviderListResponse",
        "type": "object"
      },
      "ProviderModel": {
        "description": "A model line in the detail response. Hidden and favorited models stay in the\nlist so the admin view can restore them; `GET /v1/models` is the route that\ndrops them.",
        "properties": {
          "favorite": {
            "type": "boolean"
          },
          "hidden": {
            "type": "boolean"
          },
          "id": {
            "type": "string"
          },
          "object": {
            "type": "string"
          },
          "owned_by": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "object",
          "owned_by",
          "hidden",
          "favorite"
        ],
        "type": "object"
      },
      "ProviderProtocol": {
        "description": "The wire protocol a provider speaks, as the catalog and the provider-auth\nresponses report it.\n\nThree variants, matching what the build actually serves. Node's\n`ProviderProtocol` union also lists `gemini`, but that value is dead: its only\nuser was the `gemini_cli` provider deleted with `packages/providers/src/catalog.ts`.",
        "enum": [
          "openai",
          "anthropic",
          "custom"
        ],
        "type": "string"
      },
...
      "ProviderStatus": {
        "description": "Connection state of a provider driver.",
        "properties": {
          "connected_count": {
            "format": "uint",
            "minimum": 0,
            "type": "integer"
          },
          "message": {
            "type": "string"
          },
          "state": {
            "type": "string"
          }
        },
        "required": [
          "state",
          "message",
          "connected_count"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "adminSession": {
        "description": "The admin session cookie. It also satisfies API-key auth on the routes that accept it.",
        "in": "cookie",
        "name": "srouter_admin_session",
        "type": "apiKey"
      },
      "bearerAuth": {
        "description": "The same API key in the `Authorization` header.",
        "scheme": "bearer",
        "type": "http"
      },
      "srouterApiKey": {
        "description": "An SRouter API key. `Authorization: Bearer <key>` carries the same credential.",
        "in": "header",
        "name": "x-api-key",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Multi-provider OpenAI and Anthropic compatible LLM gateway, as served by the Rust build (`server/`). One listener on `PORT` carries the API, the dashboard SPA, and the provider callbacks; there is no secondary OAuth listener. Every response uses snake_case field names (owner ruling 2026-10-07), and errors use the `{error:{message,type,code?,param?}}` envelope. Cloudflare Tunnel routes and `/v1/settings/fallbacks/*` are excluded from this build by owner ruling; they do not appear below and answer `404`.",
    "title": "SRouter API",
    "version": "0.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/models": {
      "get": {
        "parameters": [
          {
            "description": "Force a refresh of the cached catalog before answering.",
            "in": "query",
            "name": "refresh",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Force a refresh of the cached catalog before answering.",
            "in": "query",
            "name": "force",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelListResponse"
                }
              }
            },
            "description": "The allowlist-filtered catalog; hidden models and the models of disabled providers are dropped."
          }
        },
        "security": [
          {
            "srouterApiKey": []
          },
          {
            "bearerAuth": []
          },
          {
            "adminSession": []
          }
        ],
        "summary": "List the model catalog"
      },
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "favorite": {
                    "type": "boolean"
                  },
                  "hidden": {
                    "type": "boolean"
                  },
                  "model_id": {
                    "type": "string"
                  }
                },
                "required": [
                  "model_id"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogModel"
                }
              }
            },
            "description": "The model was already registered; the entry is returned."
          },
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogModel"
                }
              }
            },
            "description": "The model was registered."
          }
        },
        "security": [
          {
            "adminSession": []
          }
        ],
        "summary": "Register a custom model"
      }
    },
...
    "/v1/providers": {
      "get": {
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderListResponse"
                }
              }
            },
            "description": "One entry per built-in driver."
          }
        },
        "security": [
          {
            "srouterApiKey": []
          },
          {
            "bearerAuth": []
          },
          {
            "adminSession": []
          }
        ],
        "summary": "List the providers this build serves"
      }
    },
    "/v1/providers/catalog": {
      "get": {
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogResponse"
                }
              }
            },
            "description": "The four fixed groups."
          }
        },
        "security": [
          {
            "srouterApiKey": []
          },
          {
            "bearerAuth": []
          },
          {
            "adminSession": []
          }
        ],
        "summary": "Read the provider catalog grouped by category"
      }
    },
    "/v1/providers/{provider_id}": {
      "get": {
        "parameters": [
          {
            "description": "Provider id, for example `qoder`.",
            "in": "path",
            "name": "provider_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderEntry"
                }
              }
            },
            "description": "The provider detail entry, with its connections."
          }
        },
        "security": [
          {
            "srouterApiKey": []
          },
          {
            "bearerAuth": []
          },
          {
            "adminSession": []
          }
        ],
        "summary": "Read one provider"
      },
      "patch": {
        "parameters": [
          {
            "description": "Provider id, for example `qoder`.",
            "in": "path",
            "name": "provider_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderEntry"
                }
              }
            },
            "description": "The detail entry read back after the write."
          }
        },
        "security": [
          {
            "adminSession": []
          }
        ],
        "summary": "Toggle a provider or its model flags"
      }
    },
    "/v1/providers/{provider_id}/round-robin": {
      "patch": {
        "parameters": [
          {
            "description": "Provider id, for example `qoder`.",
            "in": "path",
            "name": "provider_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderEntry"
                }
              }
            },
            "description": "The detail entry read back after the write."
          }
        },
        "security": [
          {
            "adminSession": []
          }
        ],
        "summary": "Turn account rotation on or off"
      }
    },
...
    "/v1/v1/models": {
      "get": {
        "parameters": [
          {
            "description": "Force a refresh of the cached catalog before answering.",
            "in": "query",
            "name": "refresh",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Force a refresh of the cached catalog before answering.",
            "in": "query",
            "name": "force",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelListResponse"
                }
              }
            },
            "description": "The allowlist-filtered catalog; hidden models and the models of disabled providers are dropped."
          }
        },
        "security": [
          {
            "srouterApiKey": []
          },
          {
            "bearerAuth": []
          },
          {
            "adminSession": []
          }
        ],
        "summary": "List the model catalog"
      }
    },
...
  }
}